General fixes

+++ result.php
@@ -3,12 +3,16 @@
     include "functions.php";
     $connection = new Connection;
 
-    //print_r( $_SESSION['products'] );
-
     if ( isset( $_GET['delete'] ) ) {
         $the_id = $_GET['id'];
         $connection->deleteData( "DELETE FROM info WHERE id=$the_id" );
         header( 'Location: result.php' );
+    }
+
+    if(isset($_GET['logout'])){
+      session_unset();
+      session_destroy();
+      header("Location: index.php");
     }
 
 ?>
@@ -19,22 +23,25 @@
   <strong>Success! </strong> Product added to your wishlist.
 </div>
     <h2 class="display-4 mb-3">
-        All Records
+        <?php
+          if(isset($_SESSION['logged_username'])){
+            echo 'Welcome '.$_SESSION['logged_username'];
+          }
+        ?>
     </h2>
 
-    <a id="cart_count_link" target="_blank" href="cart-page.php" class="btn btn-info mb-3">
+    <a id="cart_count_link" href="cart-page.php" class="btn btn-info mb-3">
     <i class="fas fa-shopping-cart"></i> <span class="badge badge-light" id="cart_count">
       <?php
           if ( isset( $_GET['resetcart'] ) ) {
               $_SESSION['products'] = array();
-              $products_count = count( $_SESSION['products'] );
+              $products_count       = count( $_SESSION['products'] );
               echo $products_count;
               header( "Location: result.php" );
           } else {
               $products_array = $_SESSION['products'];
               $products_count = count( $products_array );
               echo $products_count;
-
           }
       ?>
     </span>
@@ -52,8 +59,14 @@
       <th scope="col">Unit Price</th>
       <th scope="col">Stock</th>
       <th scope="col">Add to menu</th>
-      <th scope="col">Update</th>
-      <th scope="col">Delete</th>
+      <?php
+      if("admin" == $_SESSION['logged_username']){
+        ?>
+        <th scope="col">Update</th>
+        <th scope="col">Delete</th>
+        <?php
+      }
+      ?>
     </tr>
   </thead>
   <tbody>
@@ -61,17 +74,17 @@
   <?php
       $result = $connection->getAll( 'SELECT * FROM info' );
       foreach ( $result as $res ) {
-          $id = $res['id'];
+          $id          = $res['id'];
           $productName = $res['productName'];
-          $unitPrice = $res['unitPrice'];
-          $Stock = $res['Stock'];
+          $unitPrice   = $res['unitPrice'];
+          $Stock       = $res['Stock'];
       ?>
         <tr>
           <th scope="row"><?php echo $id; ?></th>
           <td><?php echo $productName; ?></td>
           <td><?php echo $unitPrice; ?></td>
           <td><?php echo $Stock; ?></td>
-          <td><a id="<?php echo $id ?>" onclick="addToCart(<?php echo $id ?>,this)" href="javascript:;" class="btn <?php if(in_array( $id, $products_array, true )){echo 'btn-secondary';}else{echo 'btn-warning';} ?> addAllert">
+          <td><a id="<?php echo $id ?>" onclick="addToCart(<?php echo $id ?>,this)" href="javascript:;" class="btn <?php if ( in_array( $id, $products_array, true ) ) {echo 'btn-secondary';} else {echo 'btn-warning';}?> addAllert">
         <?php
             $products_array = $_SESSION['products'];
                 if ( in_array( $id, $products_array, true ) ) {
@@ -82,8 +95,14 @@
                 }
             ?>
         </a></td>
-          <td><a target="_blank" href="update.php?update&id=<?php echo $id; ?>" class="btn btn-info">Update</a></td>
+        <?php
+        if("admin" == $_SESSION['logged_username']){
+          ?>
+          <td><a href="update.php?update&id=<?php echo $id; ?>" class="btn btn-info">Update</a></td>
           <td><a href="result.php?delete&id=<?php echo $id; ?>" class="btn btn-warning">Delete</a></td>
+          <?php
+        }
+        ?>
         </tr>
 
         <?php
@@ -96,12 +115,12 @@
   </tbody>
 </table>
     </div>
-    <a href="index.php" class="btn btn-primary">Add More</a>
-    <a href="getUser.php" class="btn btn-warning">View By Category</a>
+    <a href="insert-records.php" class="btn btn-primary">Add More</a>
 
     <a href="result.php?resetcart=0" class="btn btn-primary">Reset Cart</a>
+    <a href="result.php?logout=1" class="btn btn-primary">log Out</a>
 </div>
 
-<p class="d-none" id="current_session"><?php echo json_encode($_SESSION['products']); ?></p>
+<p class="d-none" id="current_session"><?php echo json_encode( $_SESSION['products'] ); ?></p>
 
 <?php include "footer.php";?>
